Identity improvements

+++ Booking/Booking/Controllers/HotelsController.cs
@@ -1,9 +1,11 @@
 ﻿using AutoMapper;
 using AutoMapper.QueryableExtensions;
+using Booking.Services;
 using Booking.Services.ControllerServices.Interfaces;
 using Booking.Services.Interfaces;
 using Booking.ViewModels.Hotel;
 using FluentValidation;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Model.Context;
@@ -18,7 +20,9 @@ public class HotelsController(
 	IValidator<CreateHotelVm> createValidator,
 	IValidator<UpdateHotelVm> updateValidator,
 	IHotelControllerService service,
-	IPaginationService<HotelVm, HotelFilterVm> pagination
+	IPaginationService<HotelVm, HotelFilterVm> pagination,
+	IScopedIdentityService scopedIdentityService,
+	IIdentityService identityService
 ) : ControllerBase {
 
 	[HttpGet]
@@ -55,11 +59,14 @@ public class HotelsController(
 	}
 
 	[HttpPost]
+	[Authorize(Roles = "Admin,User")]
 	public async Task<IActionResult> Create([FromForm] CreateHotelVm vm) {
 		var validationResult = await createValidator.ValidateAsync(vm);
 
 		if (!validationResult.IsValid)
 			return BadRequest(validationResult.Errors);
+
+		await scopedIdentityService.InitCurrentUserAsync(this);
 
 		await service.CreateAsync(vm);
 
@@ -67,7 +74,10 @@ public class HotelsController(
 	}
 
 	[HttpPut]
+	[Authorize(Roles = "Admin,User")]
 	public async Task<IActionResult> Update([FromForm] UpdateHotelVm vm) {
+		await scopedIdentityService.InitCurrentUserAsync(this);
+
 		var validationResult = await updateValidator.ValidateAsync(vm);
 
 		if (!validationResult.IsValid)
@@ -79,8 +89,19 @@ public class HotelsController(
 	}
 
 	[HttpDelete]
+	[Authorize(Roles = "Admin,User")]
 	public async Task<IActionResult> Delete(long id) {
-		await service.DeleteIfExistsAsync(id);
+		var hotel = await context.Hotels.FirstOrDefaultAsync(h => h.Id == id);
+
+		if (hotel is not null) {
+			var user = await identityService.GetCurrentUserAsync(this);
+
+			if (hotel.UserId != user.Id)
+				return BadRequest("The hotel is not own");
+
+			await service.DeleteIfExistsAsync(id);
+		}
+
 		return Ok();
 	}
 }
